restructure project, refactor client and server
