--- keva/build.gradle
@@ -1,37 +1,0 @@
-plugins {
-    id 'java'
-    id 'idea'
-    id 'application'
-    id 'com.github.johnrengelman.shadow' version '6.0.0'
-}
-
-group 'com.jinyframework'
-version '1.0-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-dependencies {
-    compileOnly 'org.projectlombok:lombok:1.18.12'
-    implementation 'com.google.code.gson:gson:2.8.6'
-
-    compile group: 'ch.qos.logback', name:'logback-classic', version: '1.0.9'
-    compile group: 'ch.qos.logback', name:'logback-core', version: '1.0.9'
-
-    annotationProcessor 'org.projectlombok:lombok:1.18.12'
-}
-
-task runServer(type: JavaExec) {
-    group = "Execution"
-    description = "Start server"
-    classpath = sourceSets.main.runtimeClasspath
-    main = "com.jinyframework.keva.Application"
-}
-
-task runClient(type: JavaExec) {
-    group = "Execution"
-    description = "Start client"
-    classpath = sourceSets.main.runtimeClasspath
-    main = "com.jinyframework.keva.Client"
-    standardInput = System.in
-}
