+++ keva/client/build.gradle
@@ -1,0 +1,35 @@
+plugins {
+    id 'java'
+    id 'idea'
+    id 'application'
+    id 'com.github.johnrengelman.shadow' version '6.0.0'
+}
+
+group 'com.jinyframework'
+version '1.0-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+dependencies {
+    compileOnly 'org.projectlombok:lombok:1.18.12'
+    implementation 'com.google.code.gson:gson:2.8.6'
+
+    compile group: 'ch.qos.logback', name:'logback-classic', version: '1.0.9'
+    compile group: 'ch.qos.logback', name:'logback-core', version: '1.0.9'
+
+    annotationProcessor 'org.projectlombok:lombok:1.18.12'
+}
+
+shadowJar {
+    minimize()
+}
+
+run{
+    standardInput = System.in
+}
+
+application {
+    mainClassName = 'com.jinyframework.keva.client.Application'
+}
+
