+++ keva/server/src/main/java/com/jinyframework/keva/server/command/CommandHandler.java
@@ -1,0 +1,8 @@
+package com.jinyframework.keva.server.command;
+
+import java.util.List;
+
+@FunctionalInterface
+public interface CommandHandler {
+    Object handle(List<String> args);
+}
