+++ keva/server/src/main/java/com/jinyframework/keva/server/command/CommandName.java
@@ -1,0 +1,9 @@
+package com.jinyframework.keva.server.command;
+
+public enum CommandName {
+    GET,
+    SET,
+    PING,
+    STAT,
+    UNSUPPORTED
+}
