+++ keva/server/src/main/java/com/jinyframework/keva/server/command/Ping.java
@@ -1,0 +1,10 @@
+package com.jinyframework.keva.server.command;
+
+import java.util.List;
+
+public class Ping implements CommandHandler{
+    @Override
+    public String handle(List<String> args) {
+        return "PONG";
+    }
+}
