+++ keva/server/src/main/java/com/jinyframework/keva/server/command/Unsupported.java
@@ -1,0 +1,10 @@
+package com.jinyframework.keva.server.command;
+
+import java.util.List;
+
+public class Unsupported implements CommandHandler{
+    @Override
+    public String handle(List<String> args) {
+        return "Unsupported command";
+    }
+}
