+++ keva/server/src/main/java/com/jinyframework/keva/server/storage/StorageService.java
@@ -1,0 +1,12 @@
+package com.jinyframework.keva.server.storage;
+
+public final class StorageService {
+    static HashMapStorage<String, String> stringStringHashMap = new HashMapStorage<>();
+
+    private StorageService() {
+    }
+
+    public static HashMapStorage<String, String> getStringStringStore() {
+        return stringStringHashMap;
+    }
+}
