--- keva/src/main/java/com/jinyframework/keva/Client.java
@@ -1,20 +1,0 @@
-package com.jinyframework.keva;
-
-import lombok.extern.slf4j.Slf4j;
-
-@Slf4j
-public final class Client {
-    public static void main(String[] args) throws Exception {
-        KevaClient client = null;
-        try {
-            client = KevaClient.builder().host("localhost").port(6767).build();
-            client.run();
-        } catch (Exception e) {
-            log.error("Error occurred while running client: ", e);
-        } finally {
-            if (client != null) {
-                client.shutdown();
-            }
-        }
-    }
-}
