--- keva/src/main/java/com/jinyframework/keva/command/CommandHandler.java
@@ -1,8 +1,0 @@
-package com.jinyframework.keva.command;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface CommandHandler {
-    Object handle(List<String> args);
-}
