--- keva/src/main/java/com/jinyframework/keva/command/CommandName.java
@@ -1,9 +1,0 @@
-package com.jinyframework.keva.command;
-
-public enum CommandName {
-    GET,
-    SET,
-    PING,
-    STAT,
-    UNSUPPORTED
-}
