--- keva/src/main/java/com/jinyframework/keva/command/Ping.java
@@ -1,10 +1,0 @@
-package com.jinyframework.keva.command;
-
-import java.util.List;
-
-public class Ping implements CommandHandler{
-    @Override
-    public String handle(List<String> args) {
-        return "PONG";
-    }
-}
