--- keva/src/main/java/com/jinyframework/keva/storage/StorageService.java
@@ -1,12 +1,0 @@
-package com.jinyframework.keva.storage;
-
-public final class StorageService {
-    static HashMapStorage<String, String> stringStringHashMap = new HashMapStorage<>();
-
-    private StorageService() {
-    }
-
-    public static HashMapStorage<String, String> getStringStringStore() {
-        return stringStringHashMap;
-    }
-}
